move things around, add javadocs
