+++ src/com/badlogic/gdx/controllers/desktop/DesktopControllerManager.java
@@ -1,6 +1,6 @@
 package com.badlogic.gdx.controllers.desktop;
 
-import com.badlogic.gdx.controllers.sdl2.SDL2ControllerManager;
+import uk.co.electronstudio.sdl2gdx.SDL2ControllerManager;
 
 public class DesktopControllerManager extends SDL2ControllerManager {
 }
