--- src/com/badlogic/gdx/controllers/sdl2/SDL2Controllers.java
@@ -1,5 +1,0 @@
-package com.badlogic.gdx.controllers.sdl2;
-
-public class SDL2Controllers {
-
-}
