+++ src/uk/co/electronstudio/sdl2gdx/SDL2Controllers.java
@@ -1,0 +1,5 @@
+package uk.co.electronstudio.sdl2gdx;
+
+public class SDL2Controllers {
+
+}
